+++ CHANGELOG.md
@@ -17,6 +17,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Medicure now keeps actions up to date with Dependabot
 
 ### Changed
+- Improved tests' parameterizing
 - Changed typer patch method to use pure patch files.
 - Improved treat tests code coverage
 - Improved extracting episode number from file names